+++ Runtime/Codeless/IAPButton.cs
@@ -1,7 +1,5 @@
 using UnityEngine.Events;
 using UnityEngine.UI;
-using System.IO;
-using System.Collections.Generic;
 
 namespace UnityEngine.Purchasing
 {